♻️ JE-118-je-116-refactor: title 컴포넌트 분리
