+++ src/components/ActivityDetail/DetailContent/index.tsx
@@ -29,7 +29,6 @@ export default function DetailContent(props: DetailContentProps) {
         category={data.category}
         rating={data.rating}
         reviewCount={data.reviewCount}
-        id={data.id}
         isSameUser={isSameUser}
       />
       <div ref={sectionRefs.description} />
